# Improve performance for session id parser