By eliminating the use of regex (and mutex), the performance shall greatly improved, see the benchmark:

*PS: I saw some application use regex in prefix, so I choose to keep that behavior.*

Before:
```
goos: darwin
goarch: amd64
pkg: github.com/igm/sockjs-go/sockjs
BenchmarkHandler_ParseSessionID          2000000              1035 ns/op             112 B/op          2 allocs/op
BenchmarkHandler_ParseSessionID-2        1000000              1001 ns/op             112 B/op          2 allocs/op
BenchmarkHandler_ParseSessionID-4        2000000               948 ns/op             112 B/op          2 allocs/op
BenchmarkHandler_ParseSessionID-8        2000000               953 ns/op             112 B/op          2 allocs/op
PASS
ok      github.com/igm/sockjs-go/sockjs 9.711s
```

After:
```
goos: darwin
goarch: amd64
pkg: github.com/igm/sockjs-go/sockjs
BenchmarkHandler_ParseSessionID         10000000               143 ns/op              48 B/op          1 allocs/op
BenchmarkHandler_ParseSessionID-2       10000000               135 ns/op              48 B/op          1 allocs/op
BenchmarkHandler_ParseSessionID-4       10000000               137 ns/op              48 B/op          1 allocs/op
BenchmarkHandler_ParseSessionID-8       10000000               138 ns/op              48 B/op          1 allocs/op
PASS
ok      github.com/igm/sockjs-go/sockjs 6.152s
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/igm/sockjs-go/67)
<!-- Reviewable:end -->